fix(cycle-deps): helpers, plugins & ledger
