+++ packages/helpers/src/utils/liquidity.ts
@@ -1,4 +1,4 @@
-import { SwapKitNumber } from "../index";
+import { SwapKitNumber } from "../modules/swapKitNumber";
 import { BaseDecimal } from "../types/chains";
 
 type ShareParams<T extends {}> = T & { liquidityUnits: string; poolUnits: string };
